requirements.txt
Django==2.2.7
djangorestframework==3.10.3
django-cors-headers==3.5.0